Reuse same object for tagged template quasis.
Spec tagged template literals pass the same object to each invocation of
the template tag function. This way that object can be used as a cache
key, for example, using a `Map`, so that parsing of the string contents
can be cached.

This patch changes the tagged template literal transform to declare a
frozen top-level array with the quasi strings, and passes that array to
each invocation. I think it's important to adhere to the spec re:
freezing, because otherwise people might try to attach properties to the
object itself as a sort of caching mechanism, which doesn't work in spec
tagged template literals.

I'm unsure if this is the _best_ way to declare a top-level variable,
but it seems to work OK :)