Parenthesize a first spread element when necessary
In most cases this is unnecessary, but in some cases it is necessary,
notably in ternaries; for `a ? b : c.concat(…)` is very different from
`(a ? b : c).concat(…)`.

I’m not clear why the insertion of the opening parenthesis had to happen
in src/program/types/CallExpression.js rather than src/utils/spread.js
when start === element.start; but if I did it inside spread(), if I used
appendLeft or prependLeft it’d end up like `.apply((Math, a).concat(…))`
instead of `.apply(Math, (a).concat(…))`, and if I used appendRight or
prependRight, nothing happened.

Fixes #177.