Rewrite ObjectExpression transpilation
The new implementation has less code, more comments, and passes more
tests.

The design of the algorithm is detailed in comments; the only major
change to note is that the output has changed to be more conservative
with object creation. For example, `x = { ...a, b }` now compiles to

  `x = Object.assign({}, a); x.b = b`

instead of

  `x = Object.assign({}, a, {b: b})`

So generated programs that previously executed without problems may be
different now, although they will hopefully still execute without
problems.